refactor(automation): keep diagnostics boundary pure
